model: Update NodeSpec
ProseMirror allows arbitrary properties on NodeSpec. It's used by
plugins such as `gapCursor`. It is also useful for controlling plugin
behaviours on a node spec level.

This change introduces `NodeSpec.make` to be used in place of
`NodeSpec.t`. `~custom` argument is now available for passing custom
fields.

The object passed to `~custom` will be "spread" into the base spec
object to allow interop with plugins like `gapCursor`